Add generator for ordered MNIST dataset
For Topologically Densified Distributions, we need training examples
with the same number of examples of each class per batch. For ease of
calculation in the layer, this change adds a function to return a
MNIST generator that returns batches consisting of slices of each class
concatenated together, and with labels returned in a tuple so that they
can be used in custom layers when making a Keras model with multiple
inputs.