WIP Add TopologicallyDenseRegularization layer
Add a layer representing the regularization loss from the Topologically
Densified Distributions paper.
WIP: decide on a unified API for the two layers - currently one
calculated the whole loss, whereas the other just calculates the
regularization part and leaves the user to create a loss that adds the
regularization in a tf.keras.layers.Lambda layer. Need to explore with
model.add_loss and decide the best way.